move modules from integration.git to git
