+++ server/odc-service/src/main/java/com/oceanbase/odc/service/git/model/GitDiff.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,16 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.integration.git.vcs;
+package com.oceanbase.odc.service.git.model;
+
+import lombok.Data;
 
 /**
  * @author: liuyizhuo.lyz
- * @date: 2024/7/29
+ * @date: 2024/7/30
  */
-public class GithubEFacadeImpl extends GithubFacadeImpl {
-    private static final String API_URL_FORMAT = "%s/api/v3/user/repos";
+@Data
+public class GitDiff {
 
-    public GithubEFacadeImpl(String host) {
-        super(String.format(API_URL_FORMAT, host));
-    }
+    private String oldPath;
+
+    private String newPath;
+
+    private FileChangeType state;
+
+    private String diff;
+
 }
