+++ server/odc-service/src/main/java/com/oceanbase/odc/service/git/model/GitRepositoryMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.integration.git.model;
+package com.oceanbase.odc.service.git.model;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.oceanbase.odc.common.crypto.TextEncryptor;
 import com.oceanbase.odc.core.shared.PreConditions;
-import com.oceanbase.odc.metadb.integration.git.GitRepositoryEntity;
+import com.oceanbase.odc.metadb.git.GitRepositoryEntity;
 import com.oceanbase.odc.service.encryption.EncryptionFacade;
 
 import lombok.NonNull;
