+++ server/odc-service/src/main/java/com/oceanbase/odc/service/git/vcs/GitlabEFacadeImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,26 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.integration.git.model;
-
-import java.util.Date;
-
-import lombok.Data;
+package com.oceanbase.odc.service.git.vcs;
 
 /**
  * @author: liuyizhuo.lyz
  * @date: 2024/7/29
  */
-@Data
-public class GitRepository extends GitProvider {
-    private Long id;
-    private Date createTime;
-    private Date updateTime;
-    private Long creatorId;
-    private Long organizationId;
-    private Long projectId;
-    private String name;
-    private String description;
-    private String sshUrl;
-    private String cloneUrl;
+public class GitlabEFacadeImpl extends GitlabFacadeImpl {
+    private static final String API_URL_FORMAT = "%s/api/v4/projects?membership=true&simple=true";
+
+    public GitlabEFacadeImpl(String host) {
+        super(String.format(API_URL_FORMAT, host));
+    }
 }
