+++ server/odc-service/src/main/java/com/oceanbase/odc/service/git/vcs/VcsFacadeMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.integration.git.vcs;
+package com.oceanbase.odc.service.git.vcs;
 
 import org.springframework.stereotype.Component;
 
-import com.oceanbase.odc.service.integration.git.model.GitProvider;
+import com.oceanbase.odc.service.git.model.GitProvider;
 
 /**
  * @author: liuyizhuo.lyz
